+++ base_layer/core/src/base_node/comms_interface/comms_request.rs
@@ -26,10 +26,15 @@ use std::{
 };
 
 use serde::{Deserialize, Serialize};
-use tari_common_types::types::{Commitment, HashOutput, PrivateKey, PublicKey, Signature};
+use tari_common_types::types::{Commitment, FixedHash, HashOutput, PrivateKey, PublicKey, Signature};
 use tari_utilities::hex::Hex;
 
-use crate::{blocks::NewBlockTemplate, chain_storage::MmrTree, proof_of_work::PowAlgorithm};
+use crate::{
+    blocks::NewBlockTemplate,
+    chain_storage::MmrTree,
+    proof_of_work::PowAlgorithm,
+    transactions::transaction_components::OutputType,
+};
 
 /// A container for the parameters required for a FetchMmrState request.
 #[derive(Debug, Serialize, Deserialize)]
@@ -70,8 +75,13 @@ pub enum NodeCommsRequest {
     FetchMempoolTransactionsByExcessSigs {
         excess_sigs: Vec<PrivateKey>,
     },
-    FetchConstitutions {
-        dan_node_public_key: PublicKey,
+    FetchContractOutputsForBlock {
+        block_hash: HashOutput,
+        output_type: OutputType,
+    },
+    FetchContractOutputsByContractId {
+        contract_id: FixedHash,
+        output_type: OutputType,
     },
 }
 
@@ -122,9 +132,10 @@ impl Display for NodeCommsRequest {
             FetchMempoolTransactionsByExcessSigs { .. } => {
                 write!(f, "FetchMempoolTransactionsByExcessSigs")
             },
-            FetchConstitutions { .. } => {
+            FetchContractOutputsForBlock { .. } => {
                 write!(f, "FetchConstitutions")
             },
+            FetchContractOutputsByContractId { .. } => write!(f, "FetchContractOutputsByContractId"),
         }
     }
 }