+++ base_layer/core/src/transactions/coinbase_builder.rs
@@ -54,7 +54,7 @@ use crate::{
     },
 };
 
-#[derive(Debug, Clone, Error, PartialEq)]
+#[derive(Debug, Clone, Error, PartialEq, Eq)]
 pub enum CoinbaseBuildError {
     #[error("The block height for this coinbase transaction wasn't provided")]
     MissingBlockHeight,