+++ base_layer/core/src/transactions/transaction_components/side_chain/mod.rs
@@ -60,3 +60,6 @@ pub use committee_signatures::CommitteeSignatures;
 
 mod sidechain_features;
 pub use sidechain_features::{SideChainFeatures, SideChainFeaturesBuilder};
+
+mod contract_checkpoint;
+pub use contract_checkpoint::ContractCheckpoint;