+++ base_layer/p2p/src/auto_update/signature.rs
@@ -54,13 +54,9 @@ impl SignedMessageVerifier {
 
     fn verify_signature(&self, signature: &pgp::StandaloneSignature, message: &str) -> Option<&pgp::SignedPublicKey> {
         // TODO: When pgp has SignedMessage support, implement that here
-        self.maintainers.iter().find_map(|pk| {
-            if signature.verify(pk, message.as_bytes()).is_ok() {
-                Some(pk)
-            } else {
-                None
-            }
-        })
+        self.maintainers
+            .iter()
+            .find(|pk| signature.verify(pk, message.as_bytes()).is_ok())
     }
 }