+++ dan_layer/core/src/workers/states/mod.rs
@@ -49,13 +49,13 @@ pub use prepare::Prepare;
 pub use starting::Starting;
 pub use synchronizing::Synchronizing;
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum ConsensusWorkerStateEvent {
     Initialized,
     Synchronized,
     BaseLayerCheckpointNotFound,
     BaseLayerAssetRegistrationNotFound,
-    BaseLayerCommitteeDefinitionNotFound,
+    BaseLayerCheckopintNotFound,
     NotPartOfCommittee,
     Errored { reason: String },
     Prepared,